requirements.txt
discord.py
python-dotenv
pytz
requests
skyfield
pandas
numpy
utide
